backend/pyproject.toml
[tool.poetry]
name = "mangakg"
version = "0.1.0"
description = "MangaKG Django Backend"
authors = ["MangaKG Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "~5.0"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"
boto3 = "^1.34.0"
psycopg2-binary = "^2.9.0"
python-dotenv = "^1.0.0"
sentry-sdk = "^2.0.0"
gunicorn = "^21.0.0"
whitenoise = "^6.6.0"
django-cors-headers = "^4.3.0"
django-storages = {extras = ["boto3"], version = "^1.14.0"}
django-filter = "^24.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.7.0"
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"